added a new callable for locatemeta rpc only
