+++ hbase-client/src/main/java/org/apache/hadoop/hbase/client/ClusterConnection.java
@@ -342,4 +342,6 @@ public interface ClusterConnection extends Connection {
    * @throws IOException if a remote or network exception occurs
    */
   int getCurrentNrHRS() throws IOException;
+
+  MasterKeepAliveConnection getKeepAliveMasterServiceForMeta() throws MasterNotRunningException;
 }
