+++ hbase-client/src/main/java/org/apache/hadoop/hbase/client/ConnectionImplementation.java
@@ -68,7 +68,6 @@ import org.apache.hadoop.hbase.ipc.RpcControllerFactory;
 import org.apache.hadoop.hbase.regionserver.RegionServerStoppedException;
 import org.apache.hadoop.hbase.security.User;
 import org.apache.hadoop.hbase.shaded.com.google.protobuf.BlockingRpcChannel;
-import org.apache.hadoop.hbase.shaded.com.google.protobuf.RpcController;
 import org.apache.hadoop.hbase.shaded.com.google.protobuf.ServiceException;
 import org.apache.hadoop.hbase.shaded.protobuf.ProtobufUtil;
 import org.apache.hadoop.hbase.shaded.protobuf.RequestConverter;
@@ -76,42 +75,6 @@ import org.apache.hadoop.hbase.shaded.protobuf.generated.AdminProtos;
 import org.apache.hadoop.hbase.shaded.protobuf.generated.ClientProtos;
 import org.apache.hadoop.hbase.shaded.protobuf.generated.ClientProtos.ClientService.BlockingInterface;
 import org.apache.hadoop.hbase.shaded.protobuf.generated.MasterProtos;
-import org.apache.hadoop.hbase.shaded.protobuf.generated.MasterProtos.DrainRegionServersRequest;
-import org.apache.hadoop.hbase.shaded.protobuf.generated.MasterProtos.DrainRegionServersResponse;
-import org.apache.hadoop.hbase.shaded.protobuf.generated.MasterProtos.IsBalancerEnabledRequest;
-import org.apache.hadoop.hbase.shaded.protobuf.generated.MasterProtos.IsBalancerEnabledResponse;
-import org.apache.hadoop.hbase.shaded.protobuf.generated.MasterProtos.IsNormalizerEnabledRequest;
-import org.apache.hadoop.hbase.shaded.protobuf.generated.MasterProtos.IsNormalizerEnabledResponse;
-import org.apache.hadoop.hbase.shaded.protobuf.generated.MasterProtos.ListDrainingRegionServersRequest;
-import org.apache.hadoop.hbase.shaded.protobuf.generated.MasterProtos.ListDrainingRegionServersResponse;
-import org.apache.hadoop.hbase.shaded.protobuf.generated.MasterProtos.LocateMetaRequest;
-import org.apache.hadoop.hbase.shaded.protobuf.generated.MasterProtos.LocateMetaResponse;
-import org.apache.hadoop.hbase.shaded.protobuf.generated.MasterProtos.NormalizeRequest;
-import org.apache.hadoop.hbase.shaded.protobuf.generated.MasterProtos.NormalizeResponse;
-import org.apache.hadoop.hbase.shaded.protobuf.generated.MasterProtos.RemoveDrainFromRegionServersRequest;
-import org.apache.hadoop.hbase.shaded.protobuf.generated.MasterProtos.RemoveDrainFromRegionServersResponse;
-import org.apache.hadoop.hbase.shaded.protobuf.generated.MasterProtos.SecurityCapabilitiesRequest;
-import org.apache.hadoop.hbase.shaded.protobuf.generated.MasterProtos.SecurityCapabilitiesResponse;
-import org.apache.hadoop.hbase.shaded.protobuf.generated.MasterProtos.SetNormalizerRunningRequest;
-import org.apache.hadoop.hbase.shaded.protobuf.generated.MasterProtos.SetNormalizerRunningResponse;
-import org.apache.hadoop.hbase.shaded.protobuf.generated.QuotaProtos.GetQuotaStatesRequest;
-import org.apache.hadoop.hbase.shaded.protobuf.generated.QuotaProtos.GetQuotaStatesResponse;
-import org.apache.hadoop.hbase.shaded.protobuf.generated.QuotaProtos.GetSpaceQuotaRegionSizesRequest;
-import org.apache.hadoop.hbase.shaded.protobuf.generated.QuotaProtos.GetSpaceQuotaRegionSizesResponse;
-import org.apache.hadoop.hbase.shaded.protobuf.generated.ReplicationProtos.AddReplicationPeerRequest;
-import org.apache.hadoop.hbase.shaded.protobuf.generated.ReplicationProtos.AddReplicationPeerResponse;
-import org.apache.hadoop.hbase.shaded.protobuf.generated.ReplicationProtos.DisableReplicationPeerRequest;
-import org.apache.hadoop.hbase.shaded.protobuf.generated.ReplicationProtos.DisableReplicationPeerResponse;
-import org.apache.hadoop.hbase.shaded.protobuf.generated.ReplicationProtos.EnableReplicationPeerRequest;
-import org.apache.hadoop.hbase.shaded.protobuf.generated.ReplicationProtos.EnableReplicationPeerResponse;
-import org.apache.hadoop.hbase.shaded.protobuf.generated.ReplicationProtos.GetReplicationPeerConfigRequest;
-import org.apache.hadoop.hbase.shaded.protobuf.generated.ReplicationProtos.GetReplicationPeerConfigResponse;
-import org.apache.hadoop.hbase.shaded.protobuf.generated.ReplicationProtos.ListReplicationPeersRequest;
-import org.apache.hadoop.hbase.shaded.protobuf.generated.ReplicationProtos.ListReplicationPeersResponse;
-import org.apache.hadoop.hbase.shaded.protobuf.generated.ReplicationProtos.RemoveReplicationPeerRequest;
-import org.apache.hadoop.hbase.shaded.protobuf.generated.ReplicationProtos.RemoveReplicationPeerResponse;
-import org.apache.hadoop.hbase.shaded.protobuf.generated.ReplicationProtos.UpdateReplicationPeerConfigRequest;
-import org.apache.hadoop.hbase.shaded.protobuf.generated.ReplicationProtos.UpdateReplicationPeerConfigResponse;
 import org.apache.hadoop.hbase.util.Bytes;
 import org.apache.hadoop.hbase.util.EnvironmentEdgeManager;
 import org.apache.hadoop.hbase.util.ExceptionUtil;
@@ -1153,7 +1116,7 @@ class ConnectionImplementation implements ClusterConnection, Closeable {
               BlockingRpcChannel channel = rpcClient.createBlockingRpcChannel(sn, user, rpcTimeout);
               return MasterProtos.MasterService.newBlockingStub(channel);
             });
-        isMasterRunning(stub);
+        if (currMasterServerName == null) isMasterRunning(stub);
         return stub;
       } finally {
         zkw.close();
@@ -1261,6 +1224,7 @@ class ConnectionImplementation implements ClusterConnection, Closeable {
   }
 
   final MasterServiceState masterServiceState = new MasterServiceState(this);
+  MasterServiceState masterServiceStateForMeta = new MasterServiceState(this);
 
   @Override
   public MasterProtos.MasterService.BlockingInterface getMaster() throws MasterNotRunningException {
@@ -1276,20 +1240,44 @@ class ConnectionImplementation implements ClusterConnection, Closeable {
   throws MasterNotRunningException {
     synchronized (masterAndZKLock) {
       if (!isKeepAliveMasterConnectedAndRunning(this.masterServiceState)) {
+        MasterServiceStubMaker stubMaker = new MasterServiceStubMaker();
+        try {
+          this.masterServiceState.stub = stubMaker.makeStub(null);
+        } catch (MasterNotRunningException ex) {
+          throw ex;
+        } catch (IOException e) {
+          // rethrow as MasterNotRunningException so that we can keep the method sig
+          throw new MasterNotRunningException(e);
+        }
+      }
+      resetMasterServiceState(this.masterServiceState);
+    }
+    // Ugly delegation just so we can add in a Close method.
+    final MasterProtos.MasterService.BlockingInterface stub = this.masterServiceState.stub;
+    return new MasterKeepAliveConnectionDelegate(masterServiceState, stub) {
+      MasterServiceState mss = masterServiceState;
+
+      @Override
+      public void close() {
+        release(this.mss);
+      }
+    };
+  }
+
+  @Override
+  public MasterKeepAliveConnection getKeepAliveMasterServiceForMeta()
+      throws MasterNotRunningException {
+    System.out.println("in keepalive for meta");
+    synchronized (masterAndZKLock) {
+      System.out.println("in keepalive for meta inside lock");
+      if (!isKeepAliveMasterConnectedAndRunning(this.masterServiceStateForMeta)) {
         String conf_master_locs = conf.get("hbase.master.all");
         if (conf_master_locs != null) {
           System.out.println("client in keepalive  " + conf_master_locs);
           String[] master_locs = conf_master_locs.split(";");
           boolean foundNonFailedMaster = false;
-          int count = 0;
           while (!foundNonFailedMaster) {
             for (String loc : master_locs) {
-              // if (count == 0) { // TO REMOVE.just for testing since first loc is active and
-              // running
-              // count++;
-              // continue;
-              // }
-              // WRITE HELPER TO PARSE
               String[] sNprops = loc.split(",");
               String hostname = sNprops[0];
               int port = Integer.parseInt(sNprops[1]);
@@ -1301,7 +1289,7 @@ class ConnectionImplementation implements ClusterConnection, Closeable {
 
               MasterServiceStubMaker stubMaker = new MasterServiceStubMaker();
               try {
-                this.masterServiceState.stub = stubMaker.makeStub(currMasterServerName);
+                this.masterServiceStateForMeta.stub = stubMaker.makeStub(currMasterServerName);
               } catch (MasterNotRunningException ex) {
                 hasExceptions = true;
                 System.out.println("in client master is not running");
@@ -1322,11 +1310,10 @@ class ConnectionImplementation implements ClusterConnection, Closeable {
           }
         }
         else {
-          System.out.println("server in keepalive");
           // master locs not provided in conf
           MasterServiceStubMaker stubMaker = new MasterServiceStubMaker();
           try {
-            this.masterServiceState.stub = stubMaker.makeStub(null);
+            this.masterServiceStateForMeta.stub = stubMaker.makeStub(null);
           } catch (MasterNotRunningException ex) {
             System.out.println("in server master is not running");
             throw ex;
@@ -1337,514 +1324,16 @@ class ConnectionImplementation implements ClusterConnection, Closeable {
 
         }
       }
-      resetMasterServiceState(this.masterServiceState);
+      resetMasterServiceState(this.masterServiceStateForMeta);
     }
     // Ugly delegation just so we can add in a Close method.
-    final MasterProtos.MasterService.BlockingInterface stub = this.masterServiceState.stub;
-    return new MasterKeepAliveConnection() {
-      MasterServiceState mss = masterServiceState;
-
-      @Override
-      public MasterProtos.AbortProcedureResponse abortProcedure(
-          RpcController controller,
-          MasterProtos.AbortProcedureRequest request) throws ServiceException {
-        return stub.abortProcedure(controller, request);
-      }
-
-      @Override
-      public MasterProtos.ListProceduresResponse listProcedures(
-          RpcController controller,
-          MasterProtos.ListProceduresRequest request) throws ServiceException {
-        return stub.listProcedures(controller, request);
-      }
-
-      @Override
-      public MasterProtos.ListLocksResponse listLocks(
-          RpcController controller,
-          MasterProtos.ListLocksRequest request) throws ServiceException {
-        return stub.listLocks(controller, request);
-      }
-
-      @Override
-      public MasterProtos.AddColumnResponse addColumn(
-          RpcController controller,
-          MasterProtos.AddColumnRequest request) throws ServiceException {
-        return stub.addColumn(controller, request);
-      }
-
-      @Override
-      public MasterProtos.DeleteColumnResponse deleteColumn(RpcController controller,
-          MasterProtos.DeleteColumnRequest request)
-      throws ServiceException {
-        return stub.deleteColumn(controller, request);
-      }
-
-      @Override
-      public MasterProtos.ModifyColumnResponse modifyColumn(RpcController controller,
-          MasterProtos.ModifyColumnRequest request)
-      throws ServiceException {
-        return stub.modifyColumn(controller, request);
-      }
-
-      @Override
-      public MasterProtos.MoveRegionResponse moveRegion(RpcController controller,
-          MasterProtos.MoveRegionRequest request) throws ServiceException {
-        return stub.moveRegion(controller, request);
-      }
-
-      @Override
-      public MasterProtos.MergeTableRegionsResponse mergeTableRegions(
-          RpcController controller, MasterProtos.MergeTableRegionsRequest request)
-          throws ServiceException {
-        return stub.mergeTableRegions(controller, request);
-      }
-
-      @Override
-      public MasterProtos.DispatchMergingRegionsResponse dispatchMergingRegions(
-          RpcController controller, MasterProtos.DispatchMergingRegionsRequest request)
-          throws ServiceException {
-        return stub.dispatchMergingRegions(controller, request);
-      }
-
-      @Override
-      public MasterProtos.AssignRegionResponse assignRegion(RpcController controller,
-          MasterProtos.AssignRegionRequest request) throws ServiceException {
-        return stub.assignRegion(controller, request);
-      }
-
-      @Override
-      public MasterProtos.UnassignRegionResponse unassignRegion(RpcController controller,
-          MasterProtos.UnassignRegionRequest request) throws ServiceException {
-        return stub.unassignRegion(controller, request);
-      }
-
-      @Override
-      public MasterProtos.OfflineRegionResponse offlineRegion(RpcController controller,
-          MasterProtos.OfflineRegionRequest request) throws ServiceException {
-        return stub.offlineRegion(controller, request);
-      }
-
-      @Override
-      public MasterProtos.SplitTableRegionResponse splitRegion(RpcController controller,
-          MasterProtos.SplitTableRegionRequest request) throws ServiceException {
-        return stub.splitRegion(controller, request);
-      }
-
-      @Override
-      public MasterProtos.DeleteTableResponse deleteTable(RpcController controller,
-          MasterProtos.DeleteTableRequest request) throws ServiceException {
-        return stub.deleteTable(controller, request);
-      }
-
-      @Override
-      public MasterProtos.TruncateTableResponse truncateTable(RpcController controller,
-          MasterProtos.TruncateTableRequest request) throws ServiceException {
-        return stub.truncateTable(controller, request);
-      }
-
-      @Override
-      public MasterProtos.EnableTableResponse enableTable(RpcController controller,
-          MasterProtos.EnableTableRequest request) throws ServiceException {
-        return stub.enableTable(controller, request);
-      }
-
-      @Override
-      public MasterProtos.DisableTableResponse disableTable(RpcController controller,
-          MasterProtos.DisableTableRequest request) throws ServiceException {
-        return stub.disableTable(controller, request);
-      }
-
-      @Override
-      public MasterProtos.ModifyTableResponse modifyTable(RpcController controller,
-          MasterProtos.ModifyTableRequest request) throws ServiceException {
-        return stub.modifyTable(controller, request);
-      }
-
-      @Override
-      public MasterProtos.CreateTableResponse createTable(RpcController controller,
-          MasterProtos.CreateTableRequest request) throws ServiceException {
-        return stub.createTable(controller, request);
-      }
-
-      @Override
-      public MasterProtos.ShutdownResponse shutdown(RpcController controller,
-          MasterProtos.ShutdownRequest request) throws ServiceException {
-        return stub.shutdown(controller, request);
-      }
-
-      @Override
-      public MasterProtos.StopMasterResponse stopMaster(RpcController controller,
-          MasterProtos.StopMasterRequest request) throws ServiceException {
-        return stub.stopMaster(controller, request);
-      }
-
-      @Override
-      public MasterProtos.IsInMaintenanceModeResponse isMasterInMaintenanceMode(
-          final RpcController controller,
-          final MasterProtos.IsInMaintenanceModeRequest request) throws ServiceException {
-        return stub.isMasterInMaintenanceMode(controller, request);
-      }
-
-      @Override
-      public MasterProtos.BalanceResponse balance(RpcController controller,
-          MasterProtos.BalanceRequest request) throws ServiceException {
-        return stub.balance(controller, request);
-      }
-
-      @Override
-      public MasterProtos.SetBalancerRunningResponse setBalancerRunning(
-          RpcController controller, MasterProtos.SetBalancerRunningRequest request)
-          throws ServiceException {
-        return stub.setBalancerRunning(controller, request);
-      }
-
-      @Override
-      public NormalizeResponse normalize(RpcController controller,
-          NormalizeRequest request) throws ServiceException {
-        return stub.normalize(controller, request);
-      }
-
-      @Override
-      public SetNormalizerRunningResponse setNormalizerRunning(
-          RpcController controller, SetNormalizerRunningRequest request)
-          throws ServiceException {
-        return stub.setNormalizerRunning(controller, request);
-      }
-
-      @Override
-      public MasterProtos.RunCatalogScanResponse runCatalogScan(RpcController controller,
-          MasterProtos.RunCatalogScanRequest request) throws ServiceException {
-        return stub.runCatalogScan(controller, request);
-      }
-
-      @Override
-      public MasterProtos.EnableCatalogJanitorResponse enableCatalogJanitor(
-          RpcController controller, MasterProtos.EnableCatalogJanitorRequest request)
-          throws ServiceException {
-        return stub.enableCatalogJanitor(controller, request);
-      }
-
-      @Override
-      public MasterProtos.IsCatalogJanitorEnabledResponse isCatalogJanitorEnabled(
-          RpcController controller, MasterProtos.IsCatalogJanitorEnabledRequest request)
-          throws ServiceException {
-        return stub.isCatalogJanitorEnabled(controller, request);
-      }
-
-      @Override
-      public MasterProtos.RunCleanerChoreResponse runCleanerChore(RpcController controller,
-          MasterProtos.RunCleanerChoreRequest request)
-          throws ServiceException {
-        return stub.runCleanerChore(controller, request);
-      }
-
-      @Override
-      public MasterProtos.SetCleanerChoreRunningResponse setCleanerChoreRunning(
-          RpcController controller, MasterProtos.SetCleanerChoreRunningRequest request)
-          throws ServiceException {
-        return stub.setCleanerChoreRunning(controller, request);
-      }
-
-      @Override
-      public MasterProtos.IsCleanerChoreEnabledResponse isCleanerChoreEnabled(
-          RpcController controller, MasterProtos.IsCleanerChoreEnabledRequest request)
-          throws ServiceException {
-        return stub.isCleanerChoreEnabled(controller, request);
-      }
-
-      @Override
-      public ClientProtos.CoprocessorServiceResponse execMasterService(
-          RpcController controller, ClientProtos.CoprocessorServiceRequest request)
-          throws ServiceException {
-        return stub.execMasterService(controller, request);
-      }
-
-      @Override
-      public MasterProtos.SnapshotResponse snapshot(RpcController controller,
-          MasterProtos.SnapshotRequest request) throws ServiceException {
-        return stub.snapshot(controller, request);
-      }
-
-      @Override
-      public MasterProtos.GetCompletedSnapshotsResponse getCompletedSnapshots(
-          RpcController controller, MasterProtos.GetCompletedSnapshotsRequest request)
-          throws ServiceException {
-        return stub.getCompletedSnapshots(controller, request);
-      }
-
-      @Override
-      public MasterProtos.DeleteSnapshotResponse deleteSnapshot(RpcController controller,
-          MasterProtos.DeleteSnapshotRequest request) throws ServiceException {
-        return stub.deleteSnapshot(controller, request);
-      }
-
-      @Override
-      public MasterProtos.IsSnapshotDoneResponse isSnapshotDone(RpcController controller,
-          MasterProtos.IsSnapshotDoneRequest request) throws ServiceException {
-        return stub.isSnapshotDone(controller, request);
-      }
-
-      @Override
-      public MasterProtos.RestoreSnapshotResponse restoreSnapshot(
-          RpcController controller, MasterProtos.RestoreSnapshotRequest request)
-          throws ServiceException {
-        return stub.restoreSnapshot(controller, request);
-      }
-
-      @Override
-      public MasterProtos.ExecProcedureResponse execProcedure(
-          RpcController controller, MasterProtos.ExecProcedureRequest request)
-          throws ServiceException {
-        return stub.execProcedure(controller, request);
-      }
-
-      @Override
-      public MasterProtos.ExecProcedureResponse execProcedureWithRet(
-          RpcController controller, MasterProtos.ExecProcedureRequest request)
-          throws ServiceException {
-        return stub.execProcedureWithRet(controller, request);
-      }
-
-      @Override
-      public MasterProtos.IsProcedureDoneResponse isProcedureDone(RpcController controller,
-          MasterProtos.IsProcedureDoneRequest request) throws ServiceException {
-        return stub.isProcedureDone(controller, request);
-      }
-
-      @Override
-      public MasterProtos.GetProcedureResultResponse getProcedureResult(RpcController controller,
-          MasterProtos.GetProcedureResultRequest request) throws ServiceException {
-        return stub.getProcedureResult(controller, request);
-      }
-
-      @Override
-      public MasterProtos.IsMasterRunningResponse isMasterRunning(
-          RpcController controller, MasterProtos.IsMasterRunningRequest request)
-          throws ServiceException {
-        return stub.isMasterRunning(controller, request);
-      }
-
-      @Override
-      public MasterProtos.ModifyNamespaceResponse modifyNamespace(RpcController controller,
-          MasterProtos.ModifyNamespaceRequest request)
-      throws ServiceException {
-        return stub.modifyNamespace(controller, request);
-      }
-
-      @Override
-      public MasterProtos.CreateNamespaceResponse createNamespace(
-          RpcController controller,
-          MasterProtos.CreateNamespaceRequest request) throws ServiceException {
-        return stub.createNamespace(controller, request);
-      }
-
-      @Override
-      public MasterProtos.DeleteNamespaceResponse deleteNamespace(
-          RpcController controller,
-          MasterProtos.DeleteNamespaceRequest request) throws ServiceException {
-        return stub.deleteNamespace(controller, request);
-      }
-
-      @Override
-      public MasterProtos.GetNamespaceDescriptorResponse getNamespaceDescriptor(
-          RpcController controller,
-          MasterProtos.GetNamespaceDescriptorRequest request) throws ServiceException {
-        return stub.getNamespaceDescriptor(controller, request);
-      }
-
-      @Override
-      public MasterProtos.ListNamespaceDescriptorsResponse listNamespaceDescriptors(
-          RpcController controller,
-          MasterProtos.ListNamespaceDescriptorsRequest request) throws ServiceException {
-        return stub.listNamespaceDescriptors(controller, request);
-      }
-
-      @Override
-      public MasterProtos.ListTableDescriptorsByNamespaceResponse listTableDescriptorsByNamespace(
-          RpcController controller, MasterProtos.ListTableDescriptorsByNamespaceRequest request)
-              throws ServiceException {
-        return stub.listTableDescriptorsByNamespace(controller, request);
-      }
-
-      @Override
-      public MasterProtos.ListTableNamesByNamespaceResponse listTableNamesByNamespace(
-          RpcController controller, MasterProtos.ListTableNamesByNamespaceRequest request)
-              throws ServiceException {
-        return stub.listTableNamesByNamespace(controller, request);
-      }
-
-      @Override
-      public MasterProtos.GetTableStateResponse getTableState(
-              RpcController controller, MasterProtos.GetTableStateRequest request)
-              throws ServiceException {
-        return stub.getTableState(controller, request);
-      }
+    final MasterProtos.MasterService.BlockingInterface stub = this.masterServiceStateForMeta.stub;
+    return new MasterKeepAliveConnectionDelegate(masterServiceStateForMeta, stub) {
+      MasterServiceState mss = masterServiceStateForMeta;
 
       @Override
       public void close() {
         release(this.mss);
-      }
-
-      @Override
-      public MasterProtos.GetSchemaAlterStatusResponse getSchemaAlterStatus(
-          RpcController controller, MasterProtos.GetSchemaAlterStatusRequest request)
-          throws ServiceException {
-        return stub.getSchemaAlterStatus(controller, request);
-      }
-
-      @Override
-      public MasterProtos.GetTableDescriptorsResponse getTableDescriptors(
-          RpcController controller, MasterProtos.GetTableDescriptorsRequest request)
-          throws ServiceException {
-        return stub.getTableDescriptors(controller, request);
-      }
-
-      @Override
-      public MasterProtos.GetTableNamesResponse getTableNames(
-          RpcController controller, MasterProtos.GetTableNamesRequest request)
-          throws ServiceException {
-        return stub.getTableNames(controller, request);
-      }
-
-      @Override
-      public MasterProtos.GetClusterStatusResponse getClusterStatus(
-          RpcController controller, MasterProtos.GetClusterStatusRequest request)
-          throws ServiceException {
-        return stub.getClusterStatus(controller, request);
-      }
-
-      @Override
-      public MasterProtos.SetQuotaResponse setQuota(
-          RpcController controller, MasterProtos.SetQuotaRequest request)
-          throws ServiceException {
-        return stub.setQuota(controller, request);
-      }
-
-      @Override
-      public MasterProtos.MajorCompactionTimestampResponse getLastMajorCompactionTimestamp(
-          RpcController controller, MasterProtos.MajorCompactionTimestampRequest request)
-          throws ServiceException {
-        return stub.getLastMajorCompactionTimestamp(controller, request);
-      }
-
-      @Override
-      public MasterProtos.MajorCompactionTimestampResponse getLastMajorCompactionTimestampForRegion(
-          RpcController controller, MasterProtos.MajorCompactionTimestampForRegionRequest request)
-          throws ServiceException {
-        return stub.getLastMajorCompactionTimestampForRegion(controller, request);
-      }
-
-      @Override
-      public IsBalancerEnabledResponse isBalancerEnabled(RpcController controller,
-          IsBalancerEnabledRequest request) throws ServiceException {
-        return stub.isBalancerEnabled(controller, request);
-      }
-
-      @Override
-      public MasterProtos.SetSplitOrMergeEnabledResponse setSplitOrMergeEnabled(
-        RpcController controller, MasterProtos.SetSplitOrMergeEnabledRequest request)
-        throws ServiceException {
-        return stub.setSplitOrMergeEnabled(controller, request);
-      }
-
-      @Override
-      public MasterProtos.IsSplitOrMergeEnabledResponse isSplitOrMergeEnabled(
-        RpcController controller, MasterProtos.IsSplitOrMergeEnabledRequest request)
-              throws ServiceException {
-        return stub.isSplitOrMergeEnabled(controller, request);
-      }
-
-      @Override
-      public IsNormalizerEnabledResponse isNormalizerEnabled(RpcController controller,
-          IsNormalizerEnabledRequest request) throws ServiceException {
-        return stub.isNormalizerEnabled(controller, request);
-      }
-
-      @Override
-      public SecurityCapabilitiesResponse getSecurityCapabilities(RpcController controller,
-          SecurityCapabilitiesRequest request) throws ServiceException {
-        return stub.getSecurityCapabilities(controller, request);
-      }
-
-      @Override
-      public AddReplicationPeerResponse addReplicationPeer(RpcController controller,
-          AddReplicationPeerRequest request) throws ServiceException {
-        return stub.addReplicationPeer(controller, request);
-      }
-
-      @Override
-      public RemoveReplicationPeerResponse removeReplicationPeer(RpcController controller,
-          RemoveReplicationPeerRequest request) throws ServiceException {
-        return stub.removeReplicationPeer(controller, request);
-      }
-
-      @Override
-      public EnableReplicationPeerResponse enableReplicationPeer(RpcController controller,
-          EnableReplicationPeerRequest request) throws ServiceException {
-        return stub.enableReplicationPeer(controller, request);
-      }
-
-      @Override
-      public DisableReplicationPeerResponse disableReplicationPeer(RpcController controller,
-          DisableReplicationPeerRequest request) throws ServiceException {
-        return stub.disableReplicationPeer(controller, request);
-      }
-
-      @Override
-      public ListDrainingRegionServersResponse listDrainingRegionServers(RpcController controller,
-          ListDrainingRegionServersRequest request) throws ServiceException {
-        return stub.listDrainingRegionServers(controller, request);
-      }
-
-      @Override
-      public DrainRegionServersResponse drainRegionServers(RpcController controller,
-          DrainRegionServersRequest request) throws ServiceException {
-        return stub.drainRegionServers(controller, request);
-      }
-
-      @Override
-      public RemoveDrainFromRegionServersResponse removeDrainFromRegionServers(
-          RpcController controller, RemoveDrainFromRegionServersRequest request)
-          throws ServiceException {
-        return stub.removeDrainFromRegionServers(controller, request);
-      }
-
-      @Override
-      public GetReplicationPeerConfigResponse getReplicationPeerConfig(RpcController controller,
-          GetReplicationPeerConfigRequest request) throws ServiceException {
-        return stub.getReplicationPeerConfig(controller, request);
-      }
-
-      @Override
-      public LocateMetaResponse locateMeta(RpcController controller, LocateMetaRequest request)
-          throws ServiceException {
-        return stub.locateMeta(controller, request);
-      }
-
-      @Override
-      public UpdateReplicationPeerConfigResponse updateReplicationPeerConfig(
-          RpcController controller, UpdateReplicationPeerConfigRequest request)
-          throws ServiceException {
-        return stub.updateReplicationPeerConfig(controller, request);
-      }
-
-      @Override
-      public ListReplicationPeersResponse listReplicationPeers(RpcController controller,
-          ListReplicationPeersRequest request) throws ServiceException {
-        return stub.listReplicationPeers(controller, request);
-      }
-
-      @Override
-      public GetSpaceQuotaRegionSizesResponse getSpaceQuotaRegionSizes(
-          RpcController controller, GetSpaceQuotaRegionSizesRequest request)
-          throws ServiceException {
-        return stub.getSpaceQuotaRegionSizes(controller, request);
-      }
-
-      @Override
-      public GetQuotaStatesResponse getQuotaStates(
-          RpcController controller, GetQuotaStatesRequest request) throws ServiceException {
-        return stub.getQuotaStates(controller, request);
       }
     };
   }
